+++ docs/changelog.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+ - Raise error if the client API takes invalid argument
+ - Raise error if the server response returned with false-status and message in payload
 
 ## [0.2.2]
 ### Fixed